requirements.txt
crewai
python-dotenv
chromadb
pandas
mistralai
langchain-mistralai
langchain-community
streamlit
PyPDF2
plotly
openpyxl
fpdf2
tenacity